fix(driver): get rid of provideShopifyConfig and createApolloConfig functions, replace them with provideShopifyApolloDriver
