--- libs/driver/shopify/src/config/public_api.ts
@@ -1,1 +1,0 @@
-export { shopifyDriverConfig } from './shopify-driver-config';
