+++ libs/driver/shopify/src/public_api.ts
@@ -1,5 +1,3 @@
 export * from './models/public_api';
 export * from './codegen/public_api';
 export * from './graphql/public_api';
-export * from './config/public_api';
-
